Forced the runtnvme.sh script to report outcome.
Due to the logging of all stdout/stdin from tnvme, nothing was being written
to the monitor. This also increased the speed of execution by 8 fold. However,
it become aggravating  to have to open the ./Logs/current file each time just
to see if the test passed. The test results are now reported for convenience.

Additionally, we want the aggravation of not seeing tnvme's output only if we
need the speed increase, i.e. we are running tests, otherwise send tnvme's
stderr/stdout to the terminal.