Updated skiptest.cfg file.
Sometime in the past, the merging of various commits in repo tnvme caused the
removal of all the test code within group GrpNVMCompareCmd. This group was
added back in a different position and thus its reference group number changed.
Thus this skiptest file needed to follow that modifcation to keep inline.