Rearchitech gInformative and GrpInformative.
Singleton gInformative will have a new method called “Reinit()” which does t
same job as GrpInformative does today, i.e. populate the object with a curre
snapshot of DUT data. Incidentally this is also the same job that the
Informative::Informative() constructor will perform to guarantee the object
is ready to use immediately.